requirements.txt
# Core RAG Framework
llama-index>=0.10.50
llama-index-vector-stores-qdrant>=0.2.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-postprocessor-flag-embedding-reranker>=0.1.0
llama-index-llms-ollama>=0.1.0
llama-index-llms-anthropic>=0.1.0

# Document Processing
python-docx>=1.1.0
openpyxl>=3.1.2

# Vector Database
qdrant-client>=1.9.0

# API Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
tenacity>=8.2.0
structlog>=24.1.0

# ML/Embedding
torch>=2.2.0
transformers>=4.40.0
sentence-transformers>=2.7.0

# Observability
prometheus-client>=0.20.0
